pyproject.toml
[tool.poetry]
name = "pid-audit"
version = "0.1.0"
description = "P&ID to graph conversion and SOP cross-referencing pipeline"
authors = ["Kartik <your@email.com>"]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.30"
pymupdf = "^1.24"
python-docx = "^1.1"
pytesseract = "^0.3"
opencv-python-headless = "^4.9"
networkx = "^3.3"
pydantic = "^2.7"
pillow = "^10.3"
python-dotenv = "^1.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
ruff = "^0.4"

[tool.poetry.scripts]
pid-audit = "pid_audit.main:main"
pid-audit-graph-ui = "pid_audit.graph_ui:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
